fix: reduce pkg/cli/session file count for CI mergeability
Move resume smoke tests into pkg/cli/session/smoke and merge durable
show/dispatches unit tests into show.go/show_test.go so make pkg-file-count
passes the 15-file package cap without changing CLI behavior.
