+++ docs/internal/processes/api-relevant-files.md
@@ -72,7 +72,7 @@ Use this map when changing the public REST contract.
 - Live session stream invalidation identity lives on `api/components/schemas/api/FactorySessionRuntime.yaml` as `streamGenerationID`, projected from `pkg/factorysessions/projection.go` via `EngineStateSnapshot.StreamGenerationID`; keep that field aligned with the session-scoped SSE handshake header `X-Factory-Session-Stream-Generation-Id` in `pkg/api/handlers_events.go` / `pkg/interfaces.FactoryEventStream` so preflight reads and stream startup expose the same current live event-history boundary.
 - Cross-surface lifecycle exposure lives in `ui/src/api/events/api.ts` (reconnect cursors), `ui/src/features/dashboard/hooks/event-stream/useFactoryEventStream.ts`, `ui/src/features/timeline/state/timeline/replaySessionLifecycle.ts`, `ui/src/features/timeline/state/timeline/replayOrchestratorProgress.ts`, `ui/src/features/timeline/state/timeline/replayDispatchLifecycle.ts`, `ui/src/features/dashboard/components/dashboard-session-lifecycle-banner.tsx`, and `pkg/cli/session/show.go`; document API/CLI/dashboard/MCP semantics in `docs/reference/sessions.md`.
 - CLI live-and-durable pause/resume lifecycle controls belong in `pkg/cli/session/lifecycle_control.go` (`Pause`, `Resume`, `LifecycleControlConfig`, `LifecycleControlRejectedError`) with operator-facing human lines in `pkg/cli/session/lifecycle_control_human.go`; wire `you session pause [session-id]` / `you session resume [session-id]` in `pkg/cli/root_work.go`. POST to `sessionpath.ScopedPath("/pause"|"/resume", sessionID)` against the same `--server` base URI as workflow status; omit `session-id` to target `~default`, and allow named live session ids plus `dur-sess-*` durable ids so the CLI matches the shared live-session lifecycle API surface. Decode `FactorySessionLifecycleControlResponse` on HTTP 200/202/409, map 404 to not-found errors, and keep unreachable-host transport failures separate from typed lifecycle outcomes. Human output distinguishes paused/resumed, already-paused/already-running, invalid-state, not-found, and unreachable-host outcomes; global `--json` emits the API-shaped lifecycle-control response unchanged. Prove accepted/no-op/invalid-state/not-found/unreachable plus default-session and named-live-session routing, human wording, command help, and JSON outcome parity in `pkg/cli/session/lifecycle_control_test.go` and `pkg/cli/root_session_test.go`; prove CLI/API equivalence on the same lifecycle-control REST surface for default live, named live, and durable sessions in `pkg/cli/session/lifecycle_control_equivalence_test.go`. Required verification for this surface: `go test ./pkg/cli/...`, `go test ./pkg/api/servertests/... -run LifecycleControl`, and the broader repository lane `make test`.
-- Durable Factory Session dispatch inspection over the shared HTTP session command surface belongs in `pkg/cli/session/dispatches.go` (`Dispatches`, `DispatchesConfig`) with wiring in `pkg/cli/root_work.go` (`you session dispatches [session-id]`). GET `sessionpath.ScopedPath("/dispatches", sessionID)` against the same `--server` base URI as `session show` and `session resume`; require `dur-sess-*` ids, decode `ListFactorySessionDispatchesResponse`, and keep human output on FactorySession/Dispatch vocabulary. Runtime-backed resume continuity smokes in `pkg/cli/session/resume_smoke_test.go` should compare pre-resume and post-resume `session show` progress plus `session dispatches` parity to prove completed child dispatches are not replayed as new work. Runtime-backed invalid resume smokes in the same file should assert typed `TERMINAL_SESSION` / `NO_OP` lifecycle-control outcomes through `session resume` plus unchanged `session show` reads for the same `dur-sess-*` id instead of opaque transport failures. Lane `dynamic-workflows-cell-cli-resume-smoke` scope and deferrals live in `follow-up-cell-cli-resume-smoke-deferred.md`; prove additive non-resume CLI session create/show/list/lifecycle-control regression and shared-surface scope boundaries in `pkg/cli/session/resume_non_regression_test.go`. Keep generic CLI response-stream renderer work in `pkg/factorysessions/responsestream` out of scope for this lane.
+- Durable Factory Session dispatch inspection over the shared HTTP session command surface belongs in `pkg/cli/session/dispatches.go` (`Dispatches`, `DispatchesConfig`) with wiring in `pkg/cli/root_work.go` (`you session dispatches [session-id]`). GET `sessionpath.ScopedPath("/dispatches", sessionID)` against the same `--server` base URI as `session show` and `session resume`; require `dur-sess-*` ids, decode `ListFactorySessionDispatchesResponse`, and keep human output on FactorySession/Dispatch vocabulary. Durable `session show` rendering for `dur-sess-*` ids lives in `pkg/cli/session/show.go` alongside live session reads. Runtime-backed resume continuity smokes in `pkg/cli/session/smoke/resume_smoke_test.go` should compare pre-resume and post-resume `session show` progress plus `session dispatches` parity to prove completed child dispatches are not replayed as new work. Runtime-backed invalid resume smokes in the same file should assert typed `TERMINAL_SESSION` / `NO_OP` lifecycle-control outcomes through `session resume` plus unchanged `session show` reads for the same `dur-sess-*` id instead of opaque transport failures. Lane `dynamic-workflows-cell-cli-resume-smoke` scope and deferrals live in `follow-up-cell-cli-resume-smoke-deferred.md`; prove additive non-resume CLI session create/show/list/lifecycle-control regression and shared-surface scope boundaries in `pkg/cli/session/smoke/resume_non_regression_test.go`. Keep `pkg/cli/session` at or below the `make pkg-file-count` cap of 15 owned Go files by moving resume smokes into `pkg/cli/session/smoke` and merging durable show/dispatches unit tests into `show_test.go` when the parent package grows. Keep generic CLI response-stream renderer work in `pkg/factorysessions/responsestream` out of scope for this lane.
 - Orchestrator-aware public surfaces: extend `ui/src/api/factory-sessions/api.ts` with `getFactorySession` / result reads, render dashboard runtime detail through `ui/src/features/factory-session-detail/`, and expose the `factory-session` bento widget from `ui/src/features/bento/hooks/dashboardLayoutSchema.ts`.
 - Guarded dashboard bootstrap now depends on `GET /factory-sessions/{session_id}` in addition to `/factory-sessions`, `/factory-sessions/{session_id}/factory`, and `/factory-sessions/{session_id}/events`. When API contract or bootstrap sequencing changes touch that preflight route, keep browser/full-stack harnesses aligned in `ui/integration/browser-test-harness.mjs` and shared dashboard fetch doubles aligned in `ui/src/testing/app-shell-session-preflight-test-utils.ts` or the app can stall in preflight loading/error states and burn browser-lane timeouts.
 - The bounded operator proof for durable JavaScript session verification also belongs in `docs/reference/sessions.md`. When a change affects the reusable validate/status/dispatch-artifact-event/dashboard/lifecycle comparison path, update that reference topic instead of creating a separate lane-local smoke guide.
